compositor: Remove special casing of fullscreen surfaces
This was supposed to draw black borders around a fullscreen surface that
was smaller than the output.  We don't want to special case that in the
repaint loop, but may use a different shader or such.  And we want the
surface to have an opaque region that covers the output so that that
will eliminate overdraw of lower surfaces.