compositor-drm: Assign primary framebuffer in assign_planes callout
Determining whether it was possible to use a surface buffer as the
primary framebuffer was always a bit of a hack.  Just before rendering,
we look at the second top-most surface to see if it's the right size and
position and then use it if it is.  If we have a hw cursor and a drm
plane on top, it's no longer the second top-most and it wouldn't be picked
even if it could work.

With this change, we inspect surfaces in drm_assign_planes and try to find
a candidate for the primary fb there.  We track the overlap like for
overlays and if there is a surface that's not overlapped and fills the
output we pick it.

This commit also consolidate the logic to track damage as we move a surface
in and out of the primary plane, which was missing in the drm plane case.